+++ src/library/DIPS.Mobile.UI/API/Camera/BarcodeScanning/iOS/BarcodeScanner.cs
@@ -3,7 +3,7 @@ using CoreAnimation;
 using CoreFoundation;
 using CoreGraphics;
 using CoreMedia;
-using DIPS.Mobile.UI.API.Camera.iOS;
+using DIPS.Mobile.UI.API.Camera.Preview;
 using Foundation;
 using Microsoft.Maui.Controls.Shapes;
 using UIKit;
@@ -24,6 +24,8 @@ public partial class BarcodeScanner
     private AVCaptureMetadataOutput? m_captureMetadataOutput;
     private AVCaptureDeviceInput? m_videoDeviceInput;
     private double m_rectOfInterestWidth;
+
+    private DispatchQueue m_metadataObjectsQueue = new(label: "metadata objects queue", attributes: new(), target: null);
 
     //https://developer.apple.com/documentation/avfoundation/capture_setup/requesting_authorization_to_capture_and_save_media#2958841
     internal partial async Task<bool> CanUseCamera()
@@ -54,7 +56,7 @@ public partial class BarcodeScanner
 
         m_captureDevice = null;
 
-        if (m_preview?.Handler is not PreviewHandler previewHandler) return;
+        if (m_cameraPreview?.Handler is not CameraPreviewHandler previewHandler) return;
         previewHandler.RemoveZoomSlider();
         previewHandler.RemovePinchToZoom();
         previewHandler.RemoveTouchToFocus();
@@ -65,7 +67,7 @@ public partial class BarcodeScanner
     internal async partial Task PlatformStart()
     {
         //This makes sure we display the video feed
-        if (m_preview?.Handler is not PreviewHandler previewHandler) return;
+        if (m_cameraPreview?.Handler is not CameraPreviewHandler previewHandler) return;
 
         m_previewUIView = previewHandler.PlatformView;
 
@@ -94,12 +96,11 @@ public partial class BarcodeScanner
         }
 
         //Set quality for best performance
-        //Source: https://developers.google.com/ml-kit/vision/barcode-scanning/ios#performance-tips where this is mentioned
-        //But we've followed sample code from Apple: https://developer.apple.com/documentation/avfoundation/capture_setup/avcambarcode_detecting_barcodes_and_faces
-        m_captureSession.SessionPreset = AVCaptureSession.Preset1920x1080;
+        m_captureSession.SessionPreset = AVCaptureSession.PresetHigh;
 
         //Add barcode camera output
         m_captureMetadataOutput = new AVCaptureMetadataOutput();
+        
         if (m_captureSession.CanAddOutput(m_captureMetadataOutput))
         {
             m_captureSession.AddOutput(
@@ -111,7 +112,7 @@ public partial class BarcodeScanner
                 {
                     InvokeBarcodeFound(new Barcode(s.StringValue, s.Type.ToString()));
                 }
-            }), DispatchQueue.MainQueue);
+            }), m_metadataObjectsQueue);
             //Add bar code scanning metadata
             //Bar codes: https://developer.apple.com/documentation/avfoundation/avmetadataobjecttype#3801359
             //2D codes: https://developer.apple.com/documentation/avfoundation/avmetadataobjecttype#3801360
@@ -264,7 +265,7 @@ public class CaptureDelegate : AVCaptureMetadataOutputObjectsDelegate
                 var stringValue = readableObject.StringValue;
                 if (stringValue != null)
                 {
-                    m_onSuccess.Invoke(readableObject);
+                    MainThread.BeginInvokeOnMainThread(() =>m_onSuccess.Invoke(readableObject));
                 }
             }
         }