+++ src/library/DIPS.Mobile.UI/API/Camera/PreviewHandler.cs
@@ -1,6 +1,0 @@
-namespace DIPS.Mobile.UI.API.Camera;
-
-public partial class PreviewHandler
-{
-    
-}